+++ src/day03_0.rs
@@ -10,11 +10,11 @@ lazy_static! {
 
 #[must_use]
 #[allow(clippy::missing_panics_doc)]
-pub fn day() -> usize {
+pub fn day() -> u32 {
     let mut sum = 0;
     for cap in RE.captures_iter(INPUT) {
-        let l: usize = cap.get(1).unwrap().as_str().parse().unwrap();
-        let r: usize = cap.get(2).unwrap().as_str().parse().unwrap();
+        let l: u32 = cap.get(1).unwrap().as_str().parse().unwrap();
+        let r: u32 = cap.get(2).unwrap().as_str().parse().unwrap();
         sum += l * r;
     }
